Thanks to @yymao for suggestion for module-level script
